+++ client/python/vstimd/system/__init__.py
@@ -1,4 +1,5 @@
 from .system_client import SystemClient
+from .zones_models import CameraZone, CameraZoneStatus
 from .system_models import (
     Camera3D,
     InputAxisInfo,
@@ -15,6 +16,8 @@ __all__ = [
     "SystemClient",
     "CapturedFrame",
     "Camera3D",
+    "CameraZone",
+    "CameraZoneStatus",
     "InputAxisInfo",
     "InputDeviceInfo",
     "Lighting3D",